+++ test/viking/model/instance_properties/touch.js
@@ -1,6 +1,8 @@
 (function () {
     module("Viking.Model#touch", {
         setup: function() {
+            this.clock = sinon.useFakeTimers();
+            
             this.requests = [];
             this.xhr = sinon.useFakeXMLHttpRequest();
             this.xhr.onCreate = _.bind(function(xhr) {
@@ -8,37 +10,48 @@
             }, this);
         },
         teardown: function() {
+            this.clock.restore();
             this.xhr.restore();
         }
     });
 
-    test("#touch sends a PUT request to /models/:id/touch", function() {
+    test("sends a PUT request to /models/:id/touch", function() {
         var Model = Viking.Model.extend('model');
     
         var m = new Model({id: 2, foo: 'bar'});
         m.touch();
-        equal(this.requests[0].method, 'PUT');
-        equal(this.requests[0].url, '/models/2/touch');
-        deepEqual(this.requests[0].requestBody, null);
+        equal(this.requests[0].method, 'PATCH');
+        equal(this.requests[0].url, '/models/2');
+        deepEqual(JSON.parse(this.requests[0].requestBody), {
+            model: {updated_at: (new Date()).toISOString(3)}
+        });
     });
     
-    test("#touch sends an empty body", function() {
+    test("sends an empty body", function() {
         var Model = Viking.Model.extend('model');
     
         var m = new Model({id: 2, foo: 'bar'});
         m.touch();
-        deepEqual(this.requests[0].requestBody, null);
+        deepEqual(JSON.parse(this.requests[0].requestBody), {
+            model: {updated_at: (new Date()).toISOString(3)}
+        });
     });
     
-    test("#touch sends the name of the column to be touched if specified", function() {
+    test("sends the name of the column to be touched if specified", function() {
         var Model = Viking.Model.extend('model');
     
         var m = new Model({id: 2, foo: 'bar'});
         m.touch('read_at');
-        deepEqual(this.requests[0].requestBody, '{"name":"read_at"}');
+        deepEqual(JSON.parse(this.requests[0].requestBody), {
+            model: {
+                read_at: (new Date()).toISOString(3),
+                updated_at: (new Date()).toISOString(3)
+            }
+        });
+
     });
     
-    test("#touch updates the attributes from the response", function() {
+    test("updates the attributes from the response", function() {
         var Model = Viking.Model.extend('model', { coercions: {updated_at: 'Date', read_at: 'Date'} });
     
         var m = new Model({id: 2, updated_at: null, read_at: null});