+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Major Changes:
  - added `Viking.View.bindEl`
  - abstract models for `Viking.Model` can now be created
  - `connection` can be set to route ajax calls to a specific server
+ - `Viking.Model#touch` just updates the timestamps via `PATCH` instead
+   of a `GET` to `/model/id/touch`
   
 Minor Changes: